Update user's settings form
