+++ src/components/SetUserForm/SetUserForm.module.css
@@ -1,0 +1,39 @@
+.form {
+  display: flex;
+  flex-direction: column;
+}
+
+.label {
+  font-size: 1.25rem;
+}
+
+.input {
+  width: 20rem;
+  margin: 0.5rem 0;
+  padding: 0.5rem;
+  font-size: 1rem;
+  border-radius: 0.5rem;
+  border: 1px solid var(--brand-color-4);
+  outline: none;
+}
+
+.submit-button {
+  width: 10rem;
+  height: 3rem;
+  margin: 1rem 0;
+  cursor: pointer;
+  border: none;
+  border-radius: 0.5rem;
+  outline: none;
+  background-color: #b36969;
+  color: white;
+  font-size: 1.5rem;
+}
+
+.submit-button:hover {
+  background-color: #c78686;
+}
+
+.message {
+  font-size: 0.875rem;
+}
