+++ src/components/UserPanel/UserPanel.js
@@ -1,6 +1,4 @@
 import { signOut } from 'next-auth/react'
-import { useContext } from 'react'
-import { UserContext } from '@/providers/UserContext'
 import Link from 'next/link'
 import Image from 'next/image'
 import settingsImg from 'public/icons/settings.svg'
@@ -8,7 +6,6 @@ import signoutImg from 'public/icons/signout.svg'
 import styles from './UserPanel.module.css'
 
 export default function UserPanel() {
-  const session = useContext(UserContext)
   const handleSignOut = async () => {
     "use server"
     await signOut()
@@ -18,7 +15,7 @@ export default function UserPanel() {
     <div className={styles['user-panel']}>
       <div className={styles['user-avatar']}>
         <div className={styles.container}>
-          <Image
+          {/* <Image
             unoptimized
             loader={() => session.image}
             src={session.image}
@@ -26,10 +23,12 @@ export default function UserPanel() {
             width={80}
             height={80}
             className={styles['avatar-img']}
-          />
+          /> */}
         </div>
       </div>
-      <div className={styles['user-name']}>{session.name}</div>
+      <div className={styles['user-name']}>
+        {/* {session.name} */}
+      </div>
       <div className={styles['user-links']}>
         <Link
           href="/profile"
