+++ src/pages/profile/profile.module.css
@@ -1,39 +1,3 @@
 .title {
   margin: 0 0 1rem;
 }
-
-.form {
-  display: flex;
-  flex-direction: column;
-}
-
-.label {
-  font-size: 1.25rem;
-}
-
-.input {
-  width: 20rem;
-  margin: 0.5rem 0;
-  padding: 0.5rem;
-  font-size: 1rem;
-  border-radius: 0.5rem;
-  border: 1px solid var(--brand-color-4);
-  outline: none;
-}
-
-.submit-button {
-  width: 10rem;
-  height: 3rem;
-  margin-top: 1rem;
-  cursor: pointer;
-  border: none;
-  border-radius: 0.5rem;
-  outline: none;
-  background-color: #b36969;
-  color: white;
-  font-size: 1.5rem;
-}
-
-.submit-button:hover {
-  background-color: #c78686;
-}
