--- src/providers/index.js
@@ -1,11 +1,0 @@
-'use client'
-
-import UserContextProvider from './UserContext'
-
-export default function Providers({ children }) {
-  return(
-    <UserContextProvider>
-      {children}
-    </UserContextProvider>
-  )
-}
